+++ mmengine/testing/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) OpenMMLab. All rights reserved.
+from .compare import assert_allclose
+
+__all__ = ['assert_allclose']
